This commit represents 3 hours of work
I ran into two problems with the api. First, it was not handling OPTIONS
requests, which was causing a problem on the front end.  Second, the
update endpoint was not working at all, 500 errors with any request.
The api was fixed and this commit includes code to support adding,
saving, and editing records.